python/requirements.txt
# --- 核心后端 ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
python-multipart==0.0.12

# --- 异步网络与并发 ---
aiohttp==3.10.5
httpx==0.27.2

# --- 系统与工具 ---
psutil==7.2.2
PyYAML==6.0.3
packaging==26.0
lxml==6.0.2

# --- 基础工具 ---
requests==2.32.5

# --- GUI ---  
PySide6>=6.6
materialyoucolor>=3.0.0